clean_requirements.txt
absl-py==2.3.0
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
aqtinstall==3.2.1
asttokens==2.4.1
attrs==24.2.0
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.2
cement==2.10.14
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
chex==0.1.89
click==8.1.8
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
contourpy==1.2.0
cryptography==43.0.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.5
defusedxml==0.7.1
distlib==0.3.8
docutils==0.21.2
etils==1.12.2
executing==2.0.1
filelock==3.15.4
Flask==3.1.0
Flask-Cors==5.0.0
flax==0.10.6
focal-frequency-loss==0.3.0
fonttools==4.49.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
humanize==4.12.3
hyperlink==21.0.0
idna==3.7
imageio==2.37.0
imagesize==1.4.1
importlib_resources==6.5.2
incremental==24.7.2
inflate64==1.0.1
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.7
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
jedi==0.19.1
Jinja2==3.1.5
jmespath==1.0.1
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.15
keopscore==2.3
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpi4py==4.0.3
mpmath==1.3.0
msgpack==1.1.1
multivolumefile==0.2.3
mysql-connector-python==8.3.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
ntplib==0.4.0
numpy==1.26.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.5
orbax-checkpoint==0.11.16
packaging==23.2
pandas==2.2.0
parsel==1.9.1
parso==0.8.4
patch-ng==1.18.1
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
Protego==0.3.1
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
py7zr==0.22.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pybcj==1.0.6
pybind11==3.0.0
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.18.0
pykeops==2.3
PyMsgBox==1.0.9
pynput==1.7.7
pyOpenSSL==24.2.1
pyparsing==3.1.1
pyperclip==1.9.0
pyppmd==1.1.1
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytorch-fid==0.3.0
pytweening==1.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.0
pyzstd==0.17.0
qrcode==8.0
queuelib==1.7.0
referencing==0.36.2
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
roman-numerals-py==3.1.0
rpds-py==0.23.1
rubicon-objc==0.5.0
scikit-image==0.25.2
scipy==1.14.1
Scrapy==2.11.2
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.34.1
service-identity==24.1.0
setuptools==75.1.0
simplejson==3.20.1
smmap==5.0.2
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
tensorstore==0.1.75
text-unidecode==1.3
texttable==1.7.0
tifffile==2025.6.11
tinycss2==1.2.1
tk==0.1.0
tldextract==5.1.2
toml==0.10.2
toolz==1.0.0
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.4.1
tqdm==4.66.4
treescope==0.1.9
Twisted==24.7.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.26.3
w3lib==2.2.1
wandb==0.21.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
yt-dlp==2025.4.30
zipp==3.23.0
zope.interface==7.0.3
